add suggestions from pr review
